added validation, reduced the number of requests to database

+++ EngMan/Service/MessageService.cs
@@ -2,6 +2,8 @@
 using EngMan.Models;
 using EngMan.Repository;
 using System.Linq;
+using EngMan.Extensions;
+using System.Net.Http;
 
 namespace EngMan.Service
 {
@@ -17,21 +19,80 @@ namespace EngMan.Service
 
         public Message SendMessage(Message mes, int userId)
         {
-            return rep.SendMessage(mes, userId);
+            if (mes.Validate() && userId.Validate())
+            {
+                try
+                {
+                    var result = rep.SendMessage(mes, userId);
+                    return result;
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
 
         public IEnumerable<ReturnMessage> GetMessages(int userId)
         {
-            return rep.GetMessages(userId);
+            if (userId.Validate())
+            {
+                try
+                {
+                    var result = rep.GetMessages(userId);
+                    return result;
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
 
-        public IEnumerable<ReturnMessage> ReadMessages(IEnumerable<Message> messages, int userId) {
-            return rep.ReadMessages(messages, userId);
+        public IEnumerable<ReturnMessage> ReadMessages(IEnumerable<Message> messages, int userId)
+        {
+            if (userId.Validate() && messages != null)
+            {
+                if (messages.Count() > 0)
+                {
+                    foreach (var el in messages)
+                    {
+                        if (!el.Validate())
+                        {
+                            throw new HttpRequestException("Invalid model");
+                        }
+                    }
+                }
+                try
+                {
+                    var result = rep.ReadMessages(messages, userId);
+                    return result;
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
 
         public int DeleteMessage(int mesId, int userId)
         {
-            return rep.DeleteMessage(mesId, userId);
+            if (userId.Validate() && mesId.Validate())
+            {
+                try
+                {
+                    var result = rep.DeleteMessage(mesId, userId);
+                    return result;
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
     }
 }
