+++ EngMan/Service/UserDictionaryService.cs
@@ -1,5 +1,9 @@
 ﻿using EngMan.Models;
 using EngMan.Repository;
+using System.Collections.Generic;
+using EngMan.Extensions;
+using System.Net.Http;
+
 namespace EngMan.Service
 {
     public class UserDictionaryService: IUserDictionaryService
@@ -13,15 +17,81 @@ namespace EngMan.Service
 
         public UserDictionary Get(int id)
         {
-            return rep.GetUserDictionary(id);
+            if (id.Validate())
+            {
+                try
+                {
+                    return rep.GetUserDictionary(id);
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
         public UserWord Add(int id, UserWord word)
         {
-            return rep.AddWordToDictionary(id, word);
+            if (id.Validate() && word.Validate())
+            {
+                try
+                {
+                    return rep.AddWordToDictionary(id, word);
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
         public int Delete(int userId, int wordId)
         {
-            return rep.DeleteWordFromDictionary(userId, wordId);
+            if (userId.Validate() && wordId.Validate())
+            {
+                try
+                {
+                    return rep.DeleteWordFromDictionary(userId, wordId);
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
         }
+
+        public IEnumerable<string> GetAllCategories(int id)
+        {
+            if (id.Validate())
+            {
+                try
+                {
+                    return rep.GetAllCategories(id);
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
+        }
+
+        public UserDictionary GetUserDictionaryByCategory(int id, string category)
+        {
+            if (id.Validate() && category.Validate())
+            {
+                try
+                {
+                    return rep.GetUserDictionaryByCategory(id, category);
+                }
+                catch (System.Exception ex)
+                {
+                    throw new HttpRequestException(ex.Message);
+                }
+            }
+            throw new HttpRequestException("Invalid model");
+        }
+
     }
 }
